UIs Update
Created a new progress type Repair UI that is added to the viewport only if the player is in range of the generator and presses the right key.  Added new check variables to ensure the UIs are in the viewport only if the player character is interacting with the generator.