Final Commit
Finished the Repair UI`s progress bar. Changed collisions for the generator actor so the player character can only repair in certain locations, helping with animations if added. The Repair UI now only appears once when the player interacts with the generator, not every time the specific interaction key is pressed. Once the generator is fully repaired an emitter is spawned at the top of it like lights on lamps. Interaction and Repair UIs are no longer available after the generator is finished. Added check and progress variables so changes could be made easier and bugs can be patched faster.